refactor: VO 패키지로 분리하여 관리
VO에 해당하는 LottoPurchaseMoney 와 LottoNumber 를 위한 VO 패키지를 domain 하위에 두어 관리하도록 하였습니다.
