+++ src/main/java/lotto/domain/vo/LottoPurchaseMoney.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.vo;
 
 import java.util.Objects;
