+++ src/test/java/lotto/domain/LottoPurchaseMoneyTest.java
@@ -2,6 +2,7 @@ package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 
+import lotto.domain.vo.LottoPurchaseMoney;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
