+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
 import lotto.domain.enumeration.Rank;
+import lotto.domain.vo.LottoPurchaseMoney;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
